Add a failing test case for cart not being processed after an item is removed
